Allow running in xenstore mode for USB VM
Add a stub mode that disables the DBus rpc.  This mode just listens
to udev and xenstore watches and writes xenstore entries.

vusb-manager in usbvm cannot talk dbus.  It's not connected or
initialized.  When a device is removed it will call down through
common_device_del to usbmanager_device_remove.  We want this for
xenstore but not udev based removal with a USB VM.  For dom0 backend, we
need this for udev removal, so we cannot easily condition the send by
the call path.  Instead, just make the dbus sends conditioned on a dbus
connection being available.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>