Only auto assign devices when running in dom0
In the USB VM, we don't have a policy database setup.  Therefore we
cannot auto assign from the udev code path.  Plus since the policy
database is not initialized, we segfault.

We hoist the policy assignment check out of udev_maybe_add_device so it
can be called directly from the udev or xenstore paths.  This seperates
the code better for toostack vs. USB VM.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>